feat : 연습문제 1,2번 class 정리
