+++ lib/240313/Asset.dart
@@ -1,0 +1,5 @@
+abstract class Asset {
+  String name;
+
+  Asset({required this.name});
+}
