+++ lib/240313/Book.dart
@@ -1,0 +1,12 @@
+import 'Asset/abstract Asset.dart';
+
+class Book extends TangibleAsset {
+  String isbn;
+
+  Book({
+    required super.name,
+    required super.price,
+    required super.color,
+    required this.isbn,
+  });
+}
