+++ lib/240313/Computer.dart
@@ -1,0 +1,12 @@
+import 'Asset/abstract Asset.dart';
+
+class Computer extends TangibleAsset {
+  String makerName;
+
+  Computer({
+    required super.name,
+    required super.price,
+    required super.color,
+    required this.makerName,
+  });
+}
