+++ lib/240313/InTangibleAsset.dart
@@ -1,0 +1,10 @@
+import 'Asset.dart';
+
+abstract class intangibleAsset extends Asset {
+  String patent;
+
+  intangibleAsset({
+    required super.name,
+    required this.patent,
+  });
+}
